+++ fcts_maths.h
@@ -7,7 +7,7 @@
 
 float StepsToCm(int32_t nbSteps);
 float MmToCm(uint16_t ValeurMm);
-int16_t CmToSteps(float ValeurCm);
+int32_t CmToSteps(float ValeurCm);
 float Multi(float val1, uint16_t val2);
 int16_t Distance_to_temps(float objectif, float acceleration, float delta_t);
